+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 - The experimental option to import features on 20th level characters has been disabled due to conflicts with the Munching tool, and breaking features on imported characters.
 - Experimental Elixir implementation for Alchemist.
 - Alchemist subclass parsing improvements.
+- Subclass proficiencies will now parse correctly for classes imported via Feature Munching.
+- Update default 5e character/monster image in 5.2.
 
 # 6.6.18